requirements.txt
streamlit==1.56.0
pandas==3.0.2
matplotlib==3.10.8
scikit-learn==1.8.0
openai==2.31.0
python-dotenv